General: Rename ROCm backend to HIP backend
