+++ CMakeLists.txt
@@ -8,8 +8,8 @@ project(stdgpu VERSION 1.2.0
 
 set(STDGPU_BACKEND_CUDA "STDGPU_BACKEND_CUDA")
 set(STDGPU_BACKEND_OPENMP "STDGPU_BACKEND_OPENMP")
-set(STDGPU_BACKEND_ROCM "STDGPU_BACKEND_ROCM")
-list(APPEND STDGPU_BACKEND_OPTIONS ${STDGPU_BACKEND_CUDA} ${STDGPU_BACKEND_OPENMP} ${STDGPU_BACKEND_ROCM})
+set(STDGPU_BACKEND_HIP "STDGPU_BACKEND_HIP")
+list(APPEND STDGPU_BACKEND_OPTIONS ${STDGPU_BACKEND_CUDA} ${STDGPU_BACKEND_OPENMP} ${STDGPU_BACKEND_HIP})
 
 # STDGPU_BACKEND
 set(STDGPU_BACKEND ${STDGPU_BACKEND_CUDA} CACHE STRING "Device system backend, default: STDGPU_BACKEND_CUDA")
