+++ examples/hip/CMakeLists.txt
