+++ examples/hip/thrust_interoperability.cpp
