+++ examples/hip/thrust_towards_ranges.cpp
