+++ scripts/ci/configure_hip_release.sh
@@ -7,4 +7,4 @@ sh scripts/utils/create_empty_directory.sh build
 
 # Configure project
 # NOTE: Set C++ compiler to HCC rather than HIPCC. HIPCC does not pass CMake's compiler check since it needs a --amdgpu-target=<arch> flag.
-sh scripts/utils/configure_release.sh -DSTDGPU_BACKEND=STDGPU_BACKEND_ROCM -DSTDGPU_TREAT_WARNINGS_AS_ERRORS=ON -DCMAKE_CXX_COMPILER=hcc
+sh scripts/utils/configure_release.sh -DSTDGPU_BACKEND=STDGPU_BACKEND_HIP -DSTDGPU_TREAT_WARNINGS_AS_ERRORS=ON -DCMAKE_CXX_COMPILER=hcc
