+++ test/stdgpu/hip/atomic.cpp
